+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Keep container on error option: `--error-keep`
+### Changed
+- Do not keep containers by default, not even on error
+
 ## [0.0.10] - 2018-03-12
 ### Added
 - Coverage check